+++ src/RcppExports.cpp
@@ -6,8 +6,8 @@
 using namespace Rcpp;
 
 // wsCreate
-SEXP wsCreate(std::string uri, Rcpp::Environment robjPublic, Rcpp::Environment robjPrivate, Rcpp::CharacterVector accessLogChannels, Rcpp::CharacterVector errorLogChannels);
-RcppExport SEXP _websocket_wsCreate(SEXP uriSEXP, SEXP robjPublicSEXP, SEXP robjPrivateSEXP, SEXP accessLogChannelsSEXP, SEXP errorLogChannelsSEXP) {
+SEXP wsCreate(std::string uri, Rcpp::Environment robjPublic, Rcpp::Environment robjPrivate, Rcpp::CharacterVector accessLogChannels, Rcpp::CharacterVector errorLogChannels, int maxMessageSize);
+RcppExport SEXP _websocket_wsCreate(SEXP uriSEXP, SEXP robjPublicSEXP, SEXP robjPrivateSEXP, SEXP accessLogChannelsSEXP, SEXP errorLogChannelsSEXP, SEXP maxMessageSizeSEXP) {
 BEGIN_RCPP
     Rcpp::RObject rcpp_result_gen;
     Rcpp::RNGScope rcpp_rngScope_gen;
@@ -16,7 +16,8 @@ BEGIN_RCPP
     Rcpp::traits::input_parameter< Rcpp::Environment >::type robjPrivate(robjPrivateSEXP);
     Rcpp::traits::input_parameter< Rcpp::CharacterVector >::type accessLogChannels(accessLogChannelsSEXP);
     Rcpp::traits::input_parameter< Rcpp::CharacterVector >::type errorLogChannels(errorLogChannelsSEXP);
-    rcpp_result_gen = Rcpp::wrap(wsCreate(uri, robjPublic, robjPrivate, accessLogChannels, errorLogChannels));
+    Rcpp::traits::input_parameter< int >::type maxMessageSize(maxMessageSizeSEXP);
+    rcpp_result_gen = Rcpp::wrap(wsCreate(uri, robjPublic, robjPrivate, accessLogChannels, errorLogChannels, maxMessageSize));
     return rcpp_result_gen;
 END_RCPP
 }
@@ -154,7 +155,7 @@ END_RCPP
 }
 
 static const R_CallMethodDef CallEntries[] = {
-    {"_websocket_wsCreate", (DL_FUNC) &_websocket_wsCreate, 5},
+    {"_websocket_wsCreate", (DL_FUNC) &_websocket_wsCreate, 6},
     {"_websocket_wsAppendHeader", (DL_FUNC) &_websocket_wsAppendHeader, 3},
     {"_websocket_wsAddProtocols", (DL_FUNC) &_websocket_wsAddProtocols, 2},
     {"_websocket_wsConnect", (DL_FUNC) &_websocket_wsConnect, 1},