+++ NEWS.md
@@ -1,5 +1,7 @@
 # 1.1.0.9001
 
+* Added `maxMessageSize` argument to `WebSocket$new()`. ([#57](https://github.com/rstudio/websocket/pull/57))
+
 # 1.0.0
 
 * Initial release